Untangle agent-browser execution and presentation dispatch
